use Big decimal.

+++ backend/src/main/java/com/elshop/backend/product/model/request/ProductRequest.java
@@ -3,10 +3,13 @@ package com.elshop.backend.product.model.request;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 public record ProductRequest(
-        @NotBlank(message = "Title is required!")  String title,
-        @NotNull(message = "Price is required!") Double price,
+        @NotBlank(message = "Title is required!")
+        String title,
+        @NotNull(message = "Price is required!")
+        BigDecimal price,
         String category
 ) {
 }
